--- src/main/java/org/example/tablenow/domain/auth/oAuth/OAuthProvider.java
@@ -1,5 +1,0 @@
-package org.example.tablenow.domain.auth.oAuth;
-
-public enum OAuthProvider {
-    KAKAO, NAVER
-}
