+++ src/main/java/org/example/tablenow/domain/auth/oAuth/config/OAuthProvider.java
@@ -1,0 +1,5 @@
+package org.example.tablenow.domain.auth.oAuth.config;
+
+public enum OAuthProvider {
+    KAKAO, NAVER
+}
