+++ src/main/java/org/example/tablenow/domain/auth/oAuth/kakao/KakaoUserInfoResponse.java
@@ -1,4 +1,4 @@
-package org.example.tablenow.domain.auth.oAuth;
+package org.example.tablenow.domain.auth.oAuth.kakao;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
