+++ src/main/java/org/example/tablenow/domain/auth/service/KakaoAuthService.java
@@ -5,10 +5,10 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import org.example.tablenow.domain.auth.dto.response.TokenResponse;
-import org.example.tablenow.domain.auth.oAuth.KakaoUserInfo;
-import org.example.tablenow.domain.auth.oAuth.KakaoUserInfoResponse;
-import org.example.tablenow.domain.auth.oAuth.OAuthProperties;
-import org.example.tablenow.domain.auth.oAuth.OAuthProvider;
+import org.example.tablenow.domain.auth.oAuth.kakao.KakaoUserInfo;
+import org.example.tablenow.domain.auth.oAuth.kakao.KakaoUserInfoResponse;
+import org.example.tablenow.domain.auth.oAuth.config.OAuthProperties;
+import org.example.tablenow.domain.auth.oAuth.config.OAuthProvider;
 import org.example.tablenow.domain.user.entity.User;
 import org.example.tablenow.domain.user.enums.UserRole;
 import org.example.tablenow.domain.user.repository.UserRepository;
